# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WishShop API",
        "contact": {
            "email": "lijy91@foxmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "WishShop Server (Local)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "WishShop Server (Local)"
        },
        {
            "url": "https://wishshop-api-dev.thecode.me",
            "description": "WishShop Server (Development)"
        },
        {
            "url": "https://wishshop-api.thecode.me",
            "description": "WishShop Server (Production)"
        }
    ],
    "paths": {
...
        "/stores/{store_id}/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get all Orders",
                "operationId": "App\\Http\\Controllers\\OrderController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/orders/{order_id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get a specific order",
                "operationId": "App\\Http\\Controllers\\OrderController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "Order Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores": {
            "get": {
                "tags": [
                    "stores"
                ],
                "summary": "Get all Stores",
                "operationId": "App\\Http\\Controllers\\StoreController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}": {
            "get": {
                "tags": [
                    "stores"
                ],
                "summary": "Get a specific store",
                "operationId": "App\\Http\\Controllers\\StoreController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get a specific user",
                "operationId": "App\\Http\\Controllers\\UserController::show",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{user_id}/settings": {
            "get": {
                "tags": [
                    "user_settings"
                ],
                "summary": "Get all settings",
                "operationId": "App\\Http\\Controllers\\UserSettingController::index",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user_settings"
                ],
                "summary": "Create a setting",
                "operationId": "App\\Http\\Controllers\\UserSettingController::store",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "description": "Type",
                                        "type": "string",
                                        "enum": [
                                            "string",
                                            "number",
                                            "bool"
                                        ]
                                    },
                                    "key": {
                                        "description": "Key",
                                        "type": "string"
                                    },
                                    "value": {
                                        "description": "Value",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "jwt_auth": {
                "type": "apiKey",
                "description": "To make authenticated requests via http using the built in methods, you will need to set an authorization header as follows:\\n```\\nAuthorization: Bearer {yourtokenhere}\\n```\\n",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}